checkout, refactor page component
