+++ next.config.js
@@ -5,7 +5,8 @@ const nextConfig = {
   },
   images: {
     domains: [
-      "res.cloudinary.com"
+      "res.cloudinary.com",
+      "via.placeholder.com"
     ]
   }
 }
